feat: user can submit answer for question
This allows the user to only move to the next question in the Quiz once
they have selected an answer. Currently selecting "Next" will persist
the answer chosen, then move the quiz onto the next incomplete question.

In a future update, the user will be able to skip questions and come
back to them later in the quiz.