feat: track previously reviewed cards
This feature tracks cards previously viewed using `data-id` and Stimulus
value ids: Array. Once a card is viewed, the `ids` array is updated with the
card `id`. This allows the user to page back-and-forth through the cards
during review without having to flip each one to activate the btns
again.