refactor: allow subclasses to add cards to ids
The 'ids' array is used to store the Card object id. The base class
provides the collection, it is up to the subclases to decide when to add
the id to the collection. The reviewCard adds it after the review prompt
is presented, it is then used to prevent the card from being reviewed a
second time. I'll probably redesign this using a State pattern if the
complexity grows.