style: use arbitrary groups to activate elements
The Stimulus.js code for managing the styles on the controller elements
descendants was starting to be a mess. Tailwindcss provides some
mechanisms for handling a group with an arbitrary class, here we use
`.is-active` on the `.group` element to active styles on child elements.
NOTE: You need to register these arbitrary styles in the safelist
section.